Add beacon based sitemap selection
Implements Bluetooth low energy (BLE) support for beacons in
AbstractLocateBeacons. Based on the Android version the MainActivity
will create a LocateBeaconsTask with the belonging BLE API. To show the
beacons, a new fragment (NearRooms) is created. The shown beacons are
clickable to manage the beacon or change to the belonging sitemap
fragment. The location of a beacon is set in OpenHABWriteBeaconActivity,
like a NFC-Tag. The BluetoothStateBroadcastReceiver checks whether the
user turns off Bluetooth while scanning for beacons and hold the
Bluetoothadapter enabled until the BLE scanner is shut down. The
BeaconHandler will handle the known and founded beacons for the whole
system.