Implement dynamic GCM sender ID (#288)
* Implement dynamic GCM sender ID

Instead of hardcoding the sender ID of myopenhab.com in this otherwise
dynamic app, the sender ID, which is used to register the device to a
specific service account in Google Cloud Messaging, is now queried
dynamically using a new notification settings api endpoint on the
openhab-cloud endpoint.

When a user adds a new openHAB remote URL (in fact, whenever the openHAB
main activity is created), the app will start registering the device in
the used openhab-cloud instance and GCM. After all data checks, the app
will query the /api/v1/settings/notifications endpoint of this
openhab-cloud instance. This endpoint will return a JSONObject with
information about the notification settings to use (currently the GCM
sender ID only).

The returned sender ID will be used to register the android device in GCM
and, if this was successfull, the device will be registered in the
openHAB-cloud instance.

A fallback is implemented: If the sender ID could not be recognized (or is
not configured) in the openHAB-cloud instance, the currently hardcoded one
will be used.

A new information is provided in the openHAB info dialog to make it
possible to the user to check the configured sender ID. It will provide
the used sender ID and the information, if the device could be registered
in openHAB-cloud. The target audience are persons, who use their own
openHAB-cloud instance to check their configuration. The information, if
it is positive, doesn't guarantee, that push notifications will work, but
it gives a good indication if it should work and could be starting point
to find problems, if it doesn't work.

This change more or less depends on the following commit/PR to work:
 https://github.com/openhab/openhab-cloud/pull/89/commits/49d0a49fcee26c6c8bda07ca27a0729e8b79b68c
 https://github.com/openhab/openhab-cloud/pull/89

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>

* Fix crash at app startup with not remote url set

The notification settings object is only setup, if a remote url is set,
otherwise it returns null, which was not handled in the GCM registration
so far. This commit fixes this problem.