Do not start discovery service if only remote url is configured (#291)
Instead of requiring a local url to be configured in order to use the
remote url, the app now tries to connect using the local url first. if
this does not work (not configured or not reachable), it does the same for
the alt / remote url. And only if this fails, too, the service discovery
is started.

This makes testing of openhab-cloud locally a bit easier, and is what I
would expect as a user.