Implement a MySyncHttpClient for notification settings and okhttp3
Until now, there was no sync http client with okhttp3, which is needed by
the notification settings fetch (which requests the notification settings
from openHAB when it's used the first time, and therefore needs to wait
for this request). The new MySyncHttpClient class provides this kind of
functionality and does the same thing as MyAsyncHttpClient, but
synchronous.

The new MyHttpClient class is a base for both, the sync and async clients.