Fix crash if "No route to host" (#316)
* Fix crash if "No route to host"

If the connection to the remote host could not be established, the app
creates a fake response with the error message and status code 500 for the
notification settings sync hhtp request. However, until now, the creation
of this Fake response failed, as the Response builder requires some data,
which was not given so far.

This also refactors the MyHttpClient sslSetup method, which does not need
to get the whole context passed to it and therefore the constructors of
MyAsync and MySyncHttpClient does not need this neither. It now takes a
boolean argument, which indicates, if hostnames of a ssl cert should be
ignored or not.

Fixes #315

* Fix changed method signature

* Fix new method signature for Test

* Add more detailed information in error message (Exception class)

This also allows us to have a test case, which still checks that the
request failed for the correct reason but does not rely on the exact error
message anymore, which seems to be different in different circumstances.