Do not restore state after the app was stopped (#322)
Before this commit, the app tried to restore the fragments and widgets
opened before the app was closed (or killed by the system). However, this
more or less failed mostly and resulted in the problem, that, if the user
choosed a specific widget sub-section, the saved fragment was loaded (with
the title of the new fragment), instead of the new one.

This commit removes this logic. The app now does not try to restore the
state of the app anymore, after it was closed. The state is only
preserved, if the app comes to the foreground again, without being stopped
in the meantime.

Fixes #246