Add splash screen
This adds a splash screen in a way recommended by Google: https://www.youtube.com/watch?v=E5Xu2iNHRkk
It adds a new theme called "HABDroid.SplashTheme" which is set in the
Manifest. When the MainActivity is started, it replaces the splash theme
with the configured theme. During the startup the openhab icon is shown.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>