Open Notifications fragment when notification is selected
If a notification from the systems notification bar is selected, the app
now opens the notifications fragment, if it is available currently
(openHAB remote url with username and password is used).

It also shows an alert box with the message content of the selected
message, if the message was added to the intent (should always happen).
This will happen regardless if the notification fragment is opened or not
to show the full message to the user (again).

Fixes #379