Listen for broadcasts
This commit gives the user the ability to set up a custom broadcast
receiver that changes an item on receive.

This was designed to send Mi Band 2 button presses to the openHAB
server. This will only work if an app is installed, that send broadcasts
on button presses. Since Gadgetbridge does so, it is the default
broadcast.

I had to create a foreground service, that starts the receiver.
Otherwise it would be killed by the os. Foreground services have to show
a notification, but it shown as low priority, so it shouldnt disturb the
user.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>