Re-order connection settings and add separate user/pass for local openHAB
The user may want to connect to a local openHAB instance, which is also
protected by a username/password, which is most-likely a different set of
credentials as for the remote one. TO allow specifying different sets of
credentials for the remote and the local URL, a separate, second
username/password field is added for the local connection.

To have a cleaner and better separated preferences view, the connection
settings are also re-ordered and now grouped for the openHAB local
connection settings (URL, username, password) as well as the openHAB remote
connection settings (URL, username, password). Also, the encryption
settings were moved to it's own preferencescreen "Encryption settings".

Fixes #354

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>