Use Activity and PreferenceFragment for Preferences for styling
The current PreferenceActivity looks a bit detached from the rest f the
app. There's no header and, except from the back key on the phone, no easy
way to exit the preferences.

This commit changes the preference activity to not extend from
PreferenceActivity anymore, but instead from AppCompatActivity. A newly
created layout for the rpeferences also includes a toolbar for the
settings, which also holds a back key to leave the settings easily
(without the back key on the phone).

Also: The theme selection will now be applied on-the-fly, the user does
not need to leave the settings to see what the changed preference changed
in the app.