Move the preferences menu to the left navigation drawer
The design patterns from Google for the material design suggests to place
the settings menu into the left navigation drawer, if one exists. We've
one but the settings menu is still placed in the toolbar menu. This is now
changed with this commit.

However, the settings menu entry is now visible in both places to not
confuse people expecting the settings menu at the old place. However,
selecting the settings entry from the toolbar menu will show a toast
message explaining, that the entry was moved.