Change selection widgets to use radio buttons
Instead of just showing a list of all possible values where the current
value can not be easily identified anymore (without knowing it already
from the sitemap) is confusing when selecting the (probably new) item of
the selection.

Using radio buttons for the presentation of the values fixes this issue
and does not change the functionality at all.

Fixes #423