Change how it's checked, if a conversion is needed
The version is now saved as an integer, once the app starts (for every
start) and once an update intent is received, the saved version is checked
against the minimal supported version where no conversion would be
necessary. If the version is lower, the credentials will be migrated,
otherwise nothing happens.

This code can be extended later for other required conditional
conversions.