Move the clear image cache from the options to the settings menu
The possibility to clear the image cache is not an option to the currently
viewed sitemap or section or whatever. It also shouldn't be necessary to
have this option in such a prominent component as the options menu.
Therefore, and for preparation to remove the options menu completely, the
menu is moved to the settings (next to the icon type selection).

Fixes #478