Clear in-memory image cache correctly when clearing image cache
The in-memory-cache for WebImageCache isn't shared accross different
instances of this class. Creating a new one and calling clear on it clears
the disk cache, but not the in-memory cache of MyWebImage, so images may
still be loaded from the cache.

This is now fixed, as one WebImageCache is shared accross all usages.