Move the clear image cache from the options to the settings menu (#480)
* Move the clear image cache from the options to the settings menu

The possibility to clear the image cache is not an option to the currently
viewed sitemap or section or whatever. It also shouldn't be necessary to
have this option in such a prominent component as the options menu.
Therefore, and for preparation to remove the options menu completely, the
menu is moved to the settings (next to the icon type selection).

Fixes #478

* Implement review comments and fix failing test

The tone preference is used in the GcmIntentService of the full package,
haven't seen that and accidently removed it.

* Clear in-memory image cache correctly when clearing image cache

The in-memory-cache for WebImageCache isn't shared accross different
instances of this class. Creating a new one and calling clear on it clears
the disk cache, but not the in-memory cache of MyWebImage, so images may
still be loaded from the cache.

This is now fixed, as one WebImageCache is shared accross all usages.