Update settings (#404)
* Use Activity and PreferenceFragment for Preferences for styling

The current PreferenceActivity looks a bit detached from the rest f the
app. There's no header and, except from the back key on the phone, no easy
way to exit the preferences.

This commit changes the preference activity to not extend from
PreferenceActivity anymore, but instead from AppCompatActivity. A newly
created layout for the rpeferences also includes a toolbar for the
settings, which also holds a back key to leave the settings easily
(without the back key on the phone).

Also: The theme selection will now be applied on-the-fly, the user does
not need to leave the settings to see what the changed preference changed
in the app.

* Move the preferences menu to the left navigation drawer

The design patterns from Google for the material design suggests to place
the settings menu into the left navigation drawer, if one exists. We've
one but the settings menu is still placed in the toolbar menu. This is now
changed with this commit.

However, the settings menu entry is now visible in both places to not
confuse people expecting the settings menu at the old place. However,
selecting the settings entry from the toolbar menu will show a toast
message explaining, that the entry was moved.

* Revert unintentional changes

Most likely something I missed during merge.

* Fix newline in message

* Use the showMessageToUser method for creating a toast

* Use new drawable for settings icon and the same color as all others

* Remove second dividerItem

* Implement review comments and remove options entry for settings