Decouple sitemap load/selection from UI
In preparation for #538, it would be easier to have a central code class,
which handles the load of the possible sitemaps without any UI
interaction. This commit changes the current sitemap selection to use such
a class. The UI interaction (select sitemap dialog, e.g.) is still in the
UI class OpenHABMainActivity, the load of the data was moved to the
SitemapList class, which handles the load of the sitemap list.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>