Move "Select default sitemap" to Preferences
Instead of a menu entry "Select default sitemap", there's now a setting in
the Preferences menu, which allows the user to "claer the default sitemap"
selection, if there's already a selection. Once cleared and going back to
the main activity, the user is prompted to select a new default sitemap.

Fixes #479