Re-order connection settings and add separate user/pass for local ope… (#391)
* Re-order connection settings and add separate user/pass for local openHAB

The user may want to connect to a local openHAB instance, which is also
protected by a username/password, which is most-likely a different set of
credentials as for the remote one. TO allow specifying different sets of
credentials for the remote and the local URL, a separate, second
username/password field is added for the local connection.

To have a cleaner and better separated preferences view, the connection
settings are also re-ordered and now grouped for the openHAB local
connection settings (URL, username, password) as well as the openHAB remote
connection settings (URL, username, password). Also, the encryption
settings were moved to it's own preferencescreen "Encryption settings".

Fixes #354

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>

* Copy username/password to local username/password if updated

* Implement review comments

* Change how it's checked, if a conversion is needed

The version is now saved as an integer, once the app starts (for every
start) and once an update intent is received, the saved version is checked
against the minimal supported version where no conversion would be
necessary. If the version is lower, the credentials will be migrated,
otherwise nothing happens.

This code can be extended later for other required conditional
conversions.

* Use BuildConfig.VERSION_CODE instead of querying PackageManager

* Implement review comments

* Update the version used to check for necessary updates after an update
  was installed (aditionally to updating it in the main activity)
* Moving the dependencies of preferences to preference screens

* Replacing fragments instead of using preferencescreens

Instead of using the built-in functionality of PreferenceScreens to
switch between different screens, replace fragments by separate
fragments for each preference screen.

Also:
 * Fix the update mechanism of the username password preference, which
   worked the false way around (instead of putting the local user/pass
   to the remote one, the remote should be passed to the local)
 * Fix use of remote user/pass for AsyncHttpClient only

* Implement review comments

* Change strings for remote and local server settings

* [PATCH] Some cleanup.

* Remove line-break

* Fix merge error and local connection

* Pass changed username/password to PagerAdapter and DrawerAdapter

As well as not passing local username/password to the notification
settings sync http client, as this needs always to be the remote user.

* Restore getHttpCLient method in NotificationSettings

used in the full flavor, not seen.