Remove OpenHABTracker and OpenHABTrackerReceiver functionality
The discovery service was merged into the OpenHABMainActivity.

This commit also takes over things from PR #496:
* Demo mode is disabled by default and only enable it if no urls are
configured and bonjour fails. Fixes #179
* Save discovered url in settings. Fixes #11 and #248
* Inform the user that the app tries to discover oh and if that failes
it informs the user that the demo mode will be enabled. Fixes #8
* Use https for demo mode
* Discover http oh, not https. I think most people just use the default
certificate. Certificate warnings have to be turned off to make this
working, so this
adds no security at all.
* Only discover when no url is set.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>
Signed-off-by: FlorianSW <florian.schmidt.welzow@t-online.de>
Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>