Use relative URLs for http clients
Instead of always passing the full, absolute URL to the http clients
get/post/method methods, it now allows you to take a relative URL, too,
which will be evaluated against a base url, which is the openHAB URL of
the Connection object the http client was obtained from.

This makes it easier to request resources from the server of a connection,
as the caller does not need to construct the full, absolute URL by
themselfs anymore.

It's still possible to pass an absolute URL to the http client, which
will, in fact, overwrite the absolute URL, that would be based on the
openHAB base url.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>