Make instrumentation tests more independant from each other
Some tests may require some additional or specific preferences to be set
to specific values in order to work correctly.

To achieve that, this commit adds a Before method to both instrumentation
tests, which will set these required preferences before starting the
activity. It also removes the "go through intro" from the BasicWigetTest2,
which is already tested in the CheckIntroAndCancel test, so there's no
need to test it again. It is also out of scope of the test case.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>