Do not heavily rely on view Hierarchy in BasicWidgetTest (#592)
The hierarchy can vary between different api levels, so we shouldn't rely
on it when we don't need to. In these special cases fixed with this
commit, the wrapping FrameLayout of the view in question contains a
scrollIndicatorUp View in API level 22, but not in API level 26, which
changes the position of the view in question. As we do not care about
where in the structure the view is, as long as the other parameters are
correct (the view itself, and the position in other layouts of
thehierarchy).

There're probbaly other places, too, where the view finding can be
simplified, however, I would keep them as they're now, as long as they do
not work correctly anymore.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>