Modify instrumented tests for changes in Connection management
Especially the introactivity test is now a bit more complicated, as the
workflow now includes restarting the main activity after the intro is
finished. These intents are now catched up and stubbed with an OK result.
The restart of the main activity is done by the test code itself to get a
reference to the newly started activity and assign the idlingresource for
the progressbar on it.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>