Do not return cached Connections.TYPE_ANY for each other type
This was falsly implemented for the caching part in the ConnectionFactory.
Instead of checking, if the requested type is cached, it, before this
commit, returned a cached TYPE_ANY connection only, but that for each
type. This now was changed to return a cached connection, if the
requested connection type was cached before.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>