Use FasterXML/jackson-core for Sitemap deserialization
This allows us to remove the distinction of the openHAB version (JSON or
XML), as the distinction is made on the deserialization level. The
properties of each version is already the same, just the data format is
different.

Also: Using lombok instead of writing getters, setters and equals/hashCode
in each class manually.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>