Simplify layout and viewpager hierarchy
Instead of implementing our own ViewPager to implement features we do not
currently use (swiping pages) or where (imho) the benefit is at least
questionable, use the default ViewPager delivered with the support
library.

Also: Do not maintain a main activity layout for API level 21, where the
only difference to the one used for all other api levels is, that one
property is added. This property can be added to the main layout file and
will simply be ignored in api levels, which do not support it.

Also: Remove the main activity twopane layout file, which is unused in the
whole project.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>