Automatically rebuilt server depandant components on network change
If the network connectivity of the device changes, e.g. from WiFi to
mobile data, the app will rebuilt the widgetlist fragment, as well as the
drawer items, which may include fetching data from another server as
before. This ensures, that, when the user switches networks, a click on an
item, for example, is not sent to the false server. The item could've been
rendered from the local connection, and the user could've switched the
network, so that the command would be sent to the remote server, which
would be wrong. Rebuilding the UI components when the network was switched
prevents that.

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>