Use stronger binding for ConnectionFactory listeners.
Instead of using (local) broadcasts, use a direct listener interface
which makes ConnectionFactory aware of whether there currently are
listeners or not.
Use that knowledge to postpone updates after connectivity changes to a
time when we're running in foreground, as our reachability check might
fail when running in background due to e.g. doze mode.
Also re-trigger network check when coming back from background and we
are using or should use the local server, as its reachability could have
changed since we were in background.

Signed-off-by: Danny Baumann <dannybaumann@web.de>