Wait up to 2 seconds for connection to become available
Especially for local connections, like WiFi or maybe VPN, too, a
connection may not be "available" (~ connected to the internet) once it is
"connected". To allow a connection to do whatever is needed to do, wait up
to 2 seconds (in 10 200ms chunks) for the connection to become available.
If it is not available after this timeframe, the connection is returned as
"not reachable".

Fixes #755

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>