Use gradle witness for dependency verification (#740)
The plugin compairs the hashsums of the downloaded dependencies with
those in the verify list.
Because we have some libaries (e.g. gcm) not in all flavors, witness
throws an exception, when the dependency is not found.
The two options I see are:
1. Don't verify libs like gcm. However an attacker could manipulate
those libaries.
2. Remove the exception: https://github.com/mueller-ma/gradle-witness/commit/1d6be152dbaca688f3f6755a2710272e95c0bcec

The jar file is compiled with 1d6be152dbaca688f3f6755a2710272e95c0bcec
from the repo above.

Also add a hash sum for the gradle binary. Gradle wrapper uses this hash to check if the binary is compromised or not.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>