Refactor GCM implementation (#775)
* Refactor GCM implementation.

Move all GCM specifics to full package, and streamline implementation
according to latest code samples.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Remove gcmdummy project.

It's obsolete now that we have the dummy CloudMessagingHelper in the
foss flavor.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Refactor cloud connection generation.

Only handle cloud connection as available if its settings endpoint can
be reached.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Move cloud connection check to update thread.

Allows updating available and cloud connection at the same time, which
allows simolifying listener interface again.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Add method that allows waiting for init to ConnectionFactory.

This allows waiting for init to finish without using listeners. Use that
from GcmRegistrationService, and refactor OpenHABVoiceService back to
IntentService which is now possible using the new API.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix some code formatting issues.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Restore fix for #702.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Improve CloudConnection generation.

Make it share its HTTP client with the underlying remote connection, and
move fetching sender ID out of JSON to ConnectionFactory to keep all
JSON evaluation in one place.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix failing unit tests.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Ignore initial connectivity change broadcast.

We're only interested in changes, not the initial sticky broadcast.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix deadlock when using demo connection.

The initialization flag wasn't set in that case.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Avoid useless connection checks in demo mode.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Implement review comments.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Move notification settings JSON parsing to CloudConnection.

Keeps URL definition and parsing code together.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Implement review comments.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix failing unit tests.

If local or remote connection are null, we still need to run the update
handler to generate the appropriate exception.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Improve GCM status reporting.

See #560.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Do available and cloud connection checks separately.

Doing so reduces startup time if local connection is to be used and
remote connection is configured additionally.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Improve GCM status printing in about activity.

- Rename item to 'Push notification status'
- Show it in FOSS build as well
- Remove sender ID from success string
- Move foss/full specific strings to their flavors

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Move CloudConnection class to full flavor.

The sender ID getter is highly GCM specific; rename the class to
GcmCloudConnection to reflect that.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Avoid race condition in cloud connection update.

Make sure scheduling GcmRegistrationService doesn't race against the
update thread of ConnectionFactory.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Move registration status to GcmCloudConnection.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Move cloud connection class back to all-flavor code.

Moving it to full flavor means it's no longer possible to query the list
of notifications in foss flavor, which is undesired.