Make absolutely sure fragment updates are deferred until onStart().
PR #789 attempted to fix this already, but was incomplete: if a
notification was clicked and the cloud connection already was resolved
at the time the activity was created (which may or may not have
happened, as activity creation in main thread and cloud connection
resolving in update thread race against each other), fragment updates
happened even though the activity wasn't started yet. Fix this by
tracking explicitly whether the activity is started, and deferring
updates until then.

Closes #853

Signed-off-by: Danny Baumann <dannybaumann@web.de>