Add note about privacy policy and anti features to F-Droid description
In the Play Store it is possible to add a link to the privacy policy as
metadata, but in F-Droid you cannot, so we have to add it to the
description.

The note about the anti features is nice to have. It explains why the
app entry contains the line "The app has features you may not like."

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>