Refactor image loading (#850)
* Rework loading images via HTTP.

Get rid of the SmartImageView library and do image fetching via async
HTTP calls instead. Replace the previous WebImageCache solution by a
combination of small memory cache for loaded bitmaps + OkHttp caching.

The new solution also provides an indicator drawable shown while content
is loading, and cleans up the handling of the various properties
attached to our image views, such as minimum height for charts and
maximum height for images.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Clean up MJPEG streaming code.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Convert MjpegStreamer to use SyncHttpClient.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Determine bitmap cache size dynamically.

Use algorithm example from [1] which should fit well enough for us. On a
xxhdpi device with 3GB of RAM, this gives us about 36MB of cache size.

[1] https://developer.android.com/topic/performance/graphics/cache-bitmap.html

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix empty image height ratio calculation.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Unconditionally apply fallback drawable.

If loading failed we don't want to keep the progress drawable (or the
previous bitmap). If no fallback was provided, nothing at all will be
shown.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Report bitmap loading failures via failure callback.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix up HTTP caching behavior.

Don't use cache by default (we shouldn't cache API responses), and force
cache usage for icons (only fetch them if not present in cache).

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Stop clearing image cache on connection changes.

A connection change e.g. also is the initial change from null on app
startup; clearing cache in that case makes the cache mostly useless.
As the full URL is used as cache key, there shouldn't be a reason to do
so anyway.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Cancel pending image loads when detaching view.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Remove debug remnant.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix image item max height constraint.

The intended constraint is '90% of content frame height', not '90% of
content frame width'.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Safeguard against parent height being not determined yet.

Signed-off-by: Danny Baumann <dannybaumann@web.de>