Allow state loss on fragment updates while still started.
onSaveInstanceState() is also called after onPause(), so attempts to
update the fragment state in that paused state also lead to
FragmentManager throwing a IllegalStateException. As we'll get another
onSaveInstanceState() call before onStop(), allow state loss for
fragment updates in that state.

One example of this occuring was when not accepting a certificate in the
MTM dialog.

Signed-off-by: Danny Baumann <dannybaumann@web.de>