Update MTM and remove SSL ignore prefs (#849)
* Update our copy of MemorizingTrustManager.

Remove our adaptions to its code in the process; only change to upstream
is the import of the R class at this point.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Style MTM UI according to our needs.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Remove 'Ignore SSL cert' and 'Ignore SSL hostname' options.

Those are inherently risky and unneeded anyway due to us using MTM.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Integrate SSL client cert howto into SSL client cert pref.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Stop exporting base URL from connection.

Enforces all HTTP requests to be routed via the exposed HttpClient
instances.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Move all OkHttpClient initialization to ConnectionFactory.

We modify only a few properties in OkHttpClient anyway:
- SSL socket factory for client cert (shared between all clients, as
  SSLContext is a singleton)
- hostname verifier (shared between all clients, as it's derived from
  the singleton MemorizingTrustManager instance)
- cache (shared between all clients, as cache instance is a singleton)

As all properties are shared between clients, also share the actual
client instance, as this allows sharing the HTTP connection thread pool.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Fix relative URLs in NFC handling.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Remove redundant method.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Remove 'Once' option from MTM.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Adapt strings to new MTM behavior.

Reflect the fact that we can't disable cert validation anymore. Also
reword 'Always' to 'Accept' as there's no 'Once' option anymore.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Try to resolve server HTTPS instead of HTTP port.

We can safely do this now that we can properly deal with self-signed
certificates.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Correctly build resolved URL.

We query the https service via mDNS now, so we need to build the URL
accordingly.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Allow state loss on fragment updates while still started.

onSaveInstanceState() is also called after onPause(), so attempts to
update the fragment state in that paused state also lead to
FragmentManager throwing a IllegalStateException. As we'll get another
onSaveInstanceState() call before onStop(), allow state loss for
fragment updates in that state.

One example of this occuring was when not accepting a certificate in the
MTM dialog.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Stop canceling HTTP calls in onSaveInstanceState().

onSaveInstanceState() is called from onPause() (at which point we don't
want to cancel the call), and we cancel the call from onStop() anyway.

Signed-off-by: Danny Baumann <dannybaumann@web.de>

* Apply correct theme color to MTM activity TaskDescription.

Signed-off-by: Danny Baumann <dannybaumann@web.de>