Fix wrong icons being loaded if views are re-used
There's a condition where widget icons are loaded incorrectly:
- Have a WidgetImageView which just loaded an icon
- Go to another page (previous views are detached)
- Go back to initial page (previous views are reattached, but not
  necessarily in the same spot)
In step 3 of that scenario, views are likely to use a cached bitmap
instead of loading them again, in which case mLastRequest isn't properly
cleared out and re-triggered later when onAttachedToWindow() is called.
As the request points to the URL of the last spot, the wrong icon is
loaded in that case. Fix this by properly clearing out the last request
when assigning a new URL.

Signed-off-by: Danny Baumann <dannybaumann@web.de>