Next take on fixing image refresh.
Don't clear out last request when detaching view, so we can reuse it
after reattaching for resuming the refresh ... and start actually doing
the resume. Also make sure to create a request for the same purpose if
the image view was freshly created and the initial load request could be
satisfied from the cache.

Signed-off-by: Danny Baumann <dannybaumann@web.de>