Save server properties for later use
When the app is offline and the user opens the settings, some settings
like chart scaling are missing and it might be confusing to have that
settings "just sometimes". Since we don't support multiple servers
(except demo server, which is excluded), saving the flags to prefs
seems save to me.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>