Prevent swipe-to-refresh if swipe isn't started at top of page. (#1260)
Current it's too easy to accidentally trigger swipe-to-refresh when
swiping to scroll the widget list upwards. Prevent that by only allowing
swipe-to-refresh if the widget list is already at the top when starting
the swipe gesture. This means if one wants to refresh while in the
middle of the list, one needs two swipe gestures.

See #1230

Signed-off-by: Danny Baumann <dannybaumann@web.de>