Improve NFC item updates
Create a new scheme that is only used for item updates. This way we can
have a background responder for NFC item updates. It uses the framework
introduced in #1137.

If an old tag is used, we show a message to the user to renew his tag.

See https://github.com/openhab/openhab-android/issues/538#issuecomment-480483543
See #13 (Shorter parameter names and no sitemap require less bytes)

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>