Only allow sitemap NFC tag writes for sitemap links.
Previously, when long-pressing a widget in the list, one could always
write an NFC tag which navigated to the list in question. Since a widget
was pressed, though, not the list, and since all other NFC tag write
actions relate to the widget, not the list, update the logic to only
offer tag writes for widgets which contain page links.

Also clean up intent creation for tag writes in the process, to make the
different actions clearer.

Signed-off-by: Danny Baumann <dannybaumann@web.de>