Improve NFC item updates (#1302)
See https://github.com/openhab/openhab-android/issues/538#issuecomment-480483543
Closes #13 Shorter parameter names and no sitemap require less bytes. The app doesn't crash anymore if the uri is too long.

* Add label and mapped value to tag for later use

If the user reads a tag, he can see the human readable item name (label)
and state (mapping). It won't be added if the tag doesn't support saving
much data.

* Route every tag through NfcReceiveActivity

This way we can use a worker to update items without opening the app. If a sitemap needs to be opened, send an intent to MainActivity.

* Add ON/OFF for switchSupport sliders to write to tags
* Only allow sitemap NFC tag writes for sitemap links.

Previously, when long-pressing a widget in the list, one could always
write an NFC tag which navigated to the list in question. Since a widget
was pressed, though, not the list, and since all other NFC tag write
actions relate to the widget, not the list, update the logic to only
offer tag writes for widgets which contain page links.

Also clean up intent creation for tag writes in the process, to make the
different actions clearer.

* Show toast for item update tags

Fixes #1334 (Fixed by moving all code for showing openHAB toasts in one place and make sure the color is passed in the correct format to Toasty)

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>
Signed-off-by: Danny Baumann <dannybaumann@web.de>